reorganized the testutil package
